webapp/requirements.txt
matplotlib==3.4.3
numpy==1.19.5
pandas==1.3.4
seaborn==0.11.2
streamlit==1.7.0
tensorflow-cpu==2.5.0
